+++ Example/ComplexLayoutViewController.swift
@@ -14,14 +14,12 @@ class ComplexLayoutViewController: UICollectionViewController, DTCollectionViewM
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        manager.configureEvents(for: CollectionContainingCell.self) { cellType, modelType in
-            manager.register(cellType)
-            manager.sizeForCell(withItem: modelType) { [weak self] _, _ in
+        manager.register(CollectionContainingCell.self) { [weak self] mapping in
+            mapping.sizeForCell { _, _ in
                 guard let layout = self?.collectionView?.collectionViewLayout as? UICollectionViewFlowLayout else { return .zero }
                 return CGSize(width: self?.collectionView?.frame.width ?? 0, height: layout.itemSize.height)
             }
         }
-        
         manager.memoryStorage.addItems([1,3,8,15])
     }